Create Status package, wire up to website
