+++ apps/_docs/content/docs/structure.mdx
@@ -35,5 +35,6 @@ It also contains the following packages:
 | `env` | Environment variables. |
 | `feature-flags` | Feature flags. |
 | `next-config` | Shared Next.js configuration. |
+| `status` | Status (uptime) component. |
 | `tailwind-config` | Shared Tailwind CSS configuration. |
 | `typescript-config` | Shared TypeScript configuration. |
