+++ apps/web/package.json
@@ -18,6 +18,7 @@
     "@repo/env": "workspace:*",
     "@repo/feature-flags": "workspace:*",
     "@repo/next-config": "workspace:*",
+    "@repo/status": "workspace:*",
     "@repo/tailwind-config": "workspace:*",
     "@sentry/nextjs": "^8.37.1",
     "date-fns": "^4.1.0",
