+++ packages/status/package.json
@@ -1,15 +1,10 @@
 {
-  "name": "@repo/analytics",
+  "name": "@repo/status",
   "version": "0.0.0",
   "private": true,
   "dependencies": {
-    "@next/third-parties": "^15.0.2",
     "@repo/env": "workspace:*",
-    "@vercel/analytics": "^1.3.2",
-    "posthog-js": "^1.181.0",
-    "posthog-node": "^4.2.1",
-    "react": "^18.3.1",
-    "server-only": "^0.0.1"
+    "react": "^18.3.1"
   },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
